fix: Refactor project_topics to project_settings.topics
